third_party/blink/web_tests/platform/generic/fast/css/media-query-dynamic-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS iframe.contentDocument.defaultView.getComputedStyle(p).textTransform is "none"
PASS iframe.contentDocument.defaultView.getComputedStyle(p).textTransform is "uppercase"
